fix more casts

+++ src/include/QuickDraw.h
@@ -392,17 +392,17 @@ struct MatchRec { GUEST_STRUCT;
 typedef Byte *BytePtr;
 
 
-#define thePort		(STARH(STARH((GUEST<GUEST<GrafPtr>*> *) (long) SYN68K_TO_US(a5))))
-#define thePortX	((*STARH((GUEST<GUEST<GrafPtr>*> *) (long) SYN68K_TO_US(a5))))
-#define white		(STARH((GUEST<BytePtr> *)(long) SYN68K_TO_US(a5)) -8)
-#define black		(STARH((GUEST<BytePtr> *)(long) SYN68K_TO_US(a5)) -16)
-#define gray		(STARH((GUEST<BytePtr> *)(long) SYN68K_TO_US(a5)) -24)
-#define ltGray		(STARH((GUEST<BytePtr> *)(long) SYN68K_TO_US(a5)) -32)
-#define dkGray		(STARH((GUEST<BytePtr> *)(long) SYN68K_TO_US(a5)) -40)
-#define arrowX		(* (Cursor  *) (STARH((GUEST<BytePtr> *)(long) SYN68K_TO_US(a5))-108))
-#define screenBitsX	(* (BitMap  *) (STARH((GUEST<BytePtr> *)(long) SYN68K_TO_US(a5))-122))
-#define randSeed	CL(* (GUEST<LONGINT> *) (STARH((GUEST<BytePtr> *)(long) SYN68K_TO_US(a5))-126))
-#define randSeedX	((* (GUEST<LONGINT> *) (STARH((GUEST<BytePtr> *)(long) SYN68K_TO_US(a5))-126)))
+#define thePort		(STARH(STARH((GUEST<GUEST<GrafPtr>*> *) SYN68K_TO_US(a5))))
+#define thePortX	((*STARH((GUEST<GUEST<GrafPtr>*> *) SYN68K_TO_US(a5))))
+#define white		(STARH((GUEST<BytePtr> *) SYN68K_TO_US(a5)) -8)
+#define black		(STARH((GUEST<BytePtr> *) SYN68K_TO_US(a5)) -16)
+#define gray		(STARH((GUEST<BytePtr> *) SYN68K_TO_US(a5)) -24)
+#define ltGray		(STARH((GUEST<BytePtr> *) SYN68K_TO_US(a5)) -32)
+#define dkGray		(STARH((GUEST<BytePtr> *) SYN68K_TO_US(a5)) -40)
+#define arrowX		(* (Cursor  *) (STARH((GUEST<BytePtr> *) SYN68K_TO_US(a5))-108))
+#define screenBitsX	(* (BitMap  *) (STARH((GUEST<BytePtr> *) SYN68K_TO_US(a5))-122))
+#define randSeed	CL(* (GUEST<LONGINT> *) (STARH((GUEST<BytePtr> *) SYN68K_TO_US(a5))-126))
+#define randSeedX	((* (GUEST<LONGINT> *) (STARH((GUEST<BytePtr> *) SYN68K_TO_US(a5))-126)))
 
 #if 0
 extern GUEST<LONGINT> RndSeed_L;
